Added a bank utilizing servlets and a mysql db
Added a bank that supports keeping a history of transactions
for different user accounts, depositing and withdrawing from
those transactions. Added support for sessions through filters
and a persistent session in the database. Added an external config
file to set up connection from and a check if a connection can
be made before the server is started.